- moved logs component logic into TransactionLogsPanel - moved transaction details into TransactionDetailsPanel - moved common components into /components folder - added getTransactionVisibleOperations helper
