+++ src/helpers/getValue/getTransactionVisibleOperations.ts
@@ -1,0 +1,14 @@
+import { TransactionType, TransactionVisibleOperationEnum } from 'types';
+
+export const getTransactionVisibleOperations = (
+  transaction: TransactionType
+) => {
+  const operations =
+    transaction?.operations?.filter((operation) =>
+      Object.values<string>(TransactionVisibleOperationEnum).includes(
+        operation.type
+      )
+    ) ?? [];
+
+  return operations;
+};
