+++ src/helpers/getValue/index.ts
@@ -21,5 +21,6 @@ export * from './getTransactionMethod';
 export * from './getTransactionStatusIconAndColor';
 export * from './getTransactionStatusText';
 export * from './getTransactionTokens';
+export * from './getTransactionVisibleOperations';
 export * from './getTrend';
 export * from './getValidLink';
