+++ src/pages/TransactionDetails/components/TransactionInfo/TransactionDetailsPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './TransactionDetailsPanel';
