+++ src/pages/TransactionDetails/components/TransactionInfo/TransactionLogsPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './TransactionLogsPanel';
