+++ src/pages/TransactionDetails/components/index.ts
@@ -1,0 +1,6 @@
+export * from './AddressDetailItem';
+export * from './EventsList';
+export * from './OperationsList';
+export * from './ScResultsList';
+export * from './ScrDetailItem';
+export * from './TransactionInfo';
